fixed aircraft routes

+++ routes/aircraft.routes.js
@@ -2,7 +2,7 @@ const express = require("express");
 const { default: mongoose } = require("mongoose");
 const router = express.Router();
 const Aircraft = require("../models/Aircraft.model");
-
+const isAuthenticated = require("../middleware/jwt.middleware");
 //READ: list of all the aircrafts
 
 router.get("/aircrafts", (req, res, next)=> {
@@ -37,9 +37,34 @@ router.post("/aircrafts", (req, res, next)=> {
 
 //UPDATE: aircrafts
 
-router.post("/aircrafts/:aircraftId", (req, res, next)=> {
+router.get("/aircrafts/:aircraftId", (req, res, next) => {
+    const { aircraftId } = req.params;
 
-    return Aircraft.findByIdAndUpdate(req.params.aircraftId, req.body, {new: true})
+    if (!mongoose.Types.ObjectId.isValid(aircraftId)) {
+        res.status(400).json({ message: 'Specified id is not valid' });
+        return;
+      }
+
+      Aircraft.findById(aircraftId)
+      .then(aircraft => res.json(aircraft))
+    .catch(err => {
+      console.log("error getting aircraft details...", err);
+      res.status(500).json({
+        message: "error getting aircraft details...",
+        error: err
+      })
+    });
+});
+
+router.put("/aircrafts/:aircraftId", (req, res, next)=> {
+    const { aircraftId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(aircraftId)) {
+        res.status(400).json({ message: 'Specified id is not valid' });
+        return;
+      }
+
+    Aircraft.findByIdAndUpdate(aircraftId, req.body, {new: true})
     .then(aircraft => {
         
         res.status(201).json( aircraft);
@@ -55,10 +80,10 @@ router.post("/aircrafts/:aircraftId", (req, res, next)=> {
 router.delete("/aircrafts/:aircraftId", (req, res, next) => {
 const {aircraftId} = req.params;
 
-    // if(mongoose.Types.ObjectId.isValid(aircraftId)){
-    //     res.status(400).json({message: "Id is not valid"});
-    //     return;
-    // }
+    if(!mongoose.Types.ObjectId.isValid(aircraftId)){
+        res.status(400).json({message: "Id is not valid"});
+        return;
+    }
 
     Aircraft.findByIdAndRemove(aircraftId)
     .then(()=>{
